refactor: split templates/data_classes
