+++ cognite_toolkit/cdf_tk/templates/data_classes/__init__.py
@@ -1,0 +1,30 @@
+from ._config_yaml import (
+    BuildConfigYAML,
+    BuildEnvironment,
+    ConfigEntry,
+    ConfigYAMLs,
+    Environment,
+    InitConfigYAML,
+    YAMLComment,
+)
+from ._migration_yaml import Change, MigrationYAML, VersionChanges
+from ._project_directory import ProjectDirectory, ProjectDirectoryInit, ProjectDirectoryUpgrade
+from ._system_yaml import SystemConfig
+
+__all__ = [
+    "InitConfigYAML",
+    "ConfigYAMLs",
+    "SystemConfig",
+    "BuildConfigYAML",
+    "Change",
+    "VersionChanges",
+    "MigrationYAML",
+    "ProjectDirectory",
+    "ProjectDirectoryUpgrade",
+    "ProjectDirectoryInit",
+    "SystemConfig",
+    "Environment",
+    "BuildEnvironment",
+    "ConfigEntry",
+    "YAMLComment",
+]
